General cleanup 3

+++ iop/sound/audsrv/src/adpcm.c
@@ -25,7 +25,7 @@
 #include <sysclib.h>
 
 #include <audsrv.h>
-#include "audsrv.h"
+#include "audsrv_internal.h"
 #include "common.h"
 #include "rpc_server.h"
 #include "spu.h"
@@ -154,7 +154,7 @@ void *audsrv_load_adpcm(u32 *buffer, int size, int id)
 		}
 
 		/* DMA from IOP to SPU2 */
-		sceSdVoiceTrans(0, SD_TRANS_WRITE | SD_TRANS_MODE_DMA, ((u8*)buffer)+16, (u8*)adpcm->spu2_addr, adpcm->size);
+		sceSdVoiceTrans(0, SD_TRANS_WRITE | SD_TRANS_MODE_DMA, ((u8*)buffer)+16, (u32*)adpcm->spu2_addr, adpcm->size);
 		sceSdVoiceTransStatus(0, 1);
 	}
